Job Search Via Linkeln
This project will help recruiters connect with the right candidates on LinkedIn at click of a button thus saving them loads of time. If used correctly it can also help sales / marketing / business development professionals in generating leads on LinkedIn.

A recruiter working on a certain requirement will have to search either by the role or the key skills, location and things they would like to exclude from the LinkedIn profile. Once the list is populated and exported to excel the robot will start sending requests to relevant profiles.

The recruiter can touch base with the candidates who accept the request to connect to check if they are looking for a new job.z